commands,git,lfs: invert RemoteRefs() tags option
As suggested by larsxschneider on PR review, we can avoid one
negation of the new Boolean argument we have added to the RemoteRefs()
function in our "git" package, in a prior commit in this PR, if we
invert the flag's sense and rename it from ignoreTags to withTags.

We also take the opportunity to update the code comments which describe
the action of the RemoteRefs() function so as to indicate that the
inclusion of tag refs in the returned list of refs is now optional.