ssh/connection.go: report error when no session
In commit 44b8801cbb57891b4c66fc3c43551802e5532bd1 of PR #5634 we
revised the Connection() method of the SSHTransfer structure in our
"ssh" package to create SSH sessions on demand, rather than simply
return pre-existing sessions created by the setConnectionCount()
method, which allowed us to reduce the total number of SSH sessions
in many cases.

As part of this change, we altered the Connection() method to return
an error if it could not create a session successfully, and we then
updated the callers of this method to check for a non-nil error.

However, the method still returned a nil error value, along with a
nil PktlineConnection value, if the requested session ID exceeded
the maximum number of sessions permitted and so no additional session
was created or returned.  However, the method's callers all now assume
a nil error value implies a non-nil PktlineConnection value.

In particular, the batchInternal() method of the SSHBatchClient structure
in the "tq" package originally checked for a nil return value from the
Connection() method, as this was the only way the Connection() method
indicated a failure to find or create an active SSH session.  Since
the change in commit 44b8801cbb57891b4c66fc3c43551802e5532bd1, the
batchInternal() method now checks only for a non-nil error return value,
which means it does not detect the case where a nil PktlineConnection
value is returned.

As a result, in situations such as those described in #5880, the
batchInternal() method may cause a Go panic when it attempts to call
the Lock() method of a nil PktlineConnection value.  One way this can
occur is after one batch transfer operation has finished and the
Wait() method of the TransferQueue structure has called the Shutdown()
method of the sshTransfer field of the TransferQueue's concreteManifest.
That closes all the SSH sessions and sets the "conn" array to nil.
But a reference to the same SSHTransfer structure is retained by the
sshTransfer field of the SSHBatchClient structure to which the
concreteManifest's batchClientAdapter field points.  When that manifest
is reused by a subsequent batch transfer operation, the batchInternal()
method of the SSHBatchClient structure attempts to retrieve a session,
receives a nil PktlineConnection value, and causes a panic by trying
to use it to run the Lock() method.

We therefore adjust the Connection() method of the SSHTransfer structure
to return a non-nil error when the requested session ID exceeds the
maximum number of sessions allowed.  This also covers the case where
the "conn" array has been set to nil, as the len() built-in function
returns zero for the length of a nil array.