t/cmd/lfs-ssh-echo.go: set socket file permissions
In commit 448b0c4e8814eca1ea2720c665038568d8d2828d of PR #5537 our
lfs-ssh-echo test helper utility was updated to examine the value of
any provided ControlMaster command-line argument and simulate the action
of the OpenSSH program when it is passed a ControlMaster argument with
a value of either "yes" or "no".  (The other values OpenSSH supports
for this option are not accepted by our lfs-ssh-echo utility at the
moment.)

When a ControlMaster command-line argument is found, and has the value
"yes", the lfs-ssh-echo program attempts to create a temporary file at
the location given by the ControlPath command-line argument, which must
also be provided.  If the file already exists, or some other error
occurs while creating it, the program halts with a non-zero exit code.

If the ControlMaster argument's value is "no", and the ControlPath
argument is also defined, the program attempts to open the file at
the given path, and exits with a non-zero code if the file does not
already exist or some other error occurs.

This logic is designed to emulate the behaviour of OpenSSH, which
supports the use of these arguments to multiplex SSH sessions over
a common connection using a control socket.  When the ControlMaster
argument is "yes", OpenSSH creates a socket, which may then be shared
with other invocations of OpenSSH by setting the ControlMaster argument
to "no" and passing the socket's path in the ControlPath argument.

Our lfs-ssh-echo program, however, creates its temporary file with
no defined file permissions (i.e., with a file mode argument of zero).
This results in a file which can not be opened by any other instances
of the lfs-ssh-echo program.  At the moment this does not cause any
problems with our test suite, because we have no tests which exercise
the SSH version of the Git LFS object transfer protocol with more
than a single object, so multiple SSH sessions with the same
ControlPath argument are never started.

We would like to create additional tests which push and fetch multiple
Git LFS objects over SSH, though, to help diagnose issues such as
those described in #5880.  Therefore we change our lfs-ssh-echo
program to use the conventional file mode of 0666 when creating its
temporary file, which will allow subsequent invocations of the program
with the ControlMaster argument set to "no" to also open the same file.

As well as fixing this bug, we also alter the way the lfs-ssh-echo
program handles errors when the file specified by the ControlPath
argument can not be created or opened.  We now output distinct error
messages in the cases where the file already exists or does not exist,
respectively, as well as in the cases where some other type of error
occurs.  Previously, we assumed any error must be due to a file either
existing or not existing, and did not report the actual error message,
so problems such as those caused by missing file permissions were
obscured.