lfsapi/lfsapi.go: enhance trace logs
Since PR #4446 the SSHTransfer() method of the Client structure in
our "lfsapi" package has output several trace log messages when
attempting to instantiate a new SSHTransfer structure.  This method
takes "operation" and "remote" string arguments, and the returned
SSHTransfer structure is specific to those values.

As this method may be called several times with distinct arguments
during the execution of a Git LFS process, we add the "operation"
and "remote" variables to the trace log messages, which will help
clarify the calling context of the method in our diagnostic logs.