script/build-git: update Ubuntu 24.04 APT sources
As of Ubuntu 24.04, the primary list of sources for APT moved to the
/etc/apt/sources.list.d/ubuntu.sources file instead of the legacy
/etc/apt/sources.list file, and was rewritten to make use of the
deb822 format, which is based on RFC 822:

  https://discourse.ubuntu.com/t/ubuntu-24-04-lts-noble-numbat-release-notes/39890#deb822-sources-management
  https://discourse.ubuntu.com/t/spec-apt-deb822-sources-by-default/29333
  https://manpages.ubuntu.com/manpages/noble/man5/sources.list.5.html
  https://manpages.debian.org/bookworm/dpkg-dev/deb822.5.en.html

As GitHub Actions are now migrating their "ubuntu-latest" runners
to Ubuntu 24.04, we need to adjust the script our CI jobs use to build
custom versions of Git so that it is able to fetch the necessary
source packages it requires.  We therefore update the script so it
checks for the presence of an /etc/apt/sources.list.d/ubuntu.sources
file, and if that exists, adds the "deb-src" archive type to the
"Types:" lines of the entries in that file; otherwise, the script
adds the "deb-src" type to the entries in the /etc/apt/sources.list
file as it did previously.