t: remove spurious sample test comments
In commit 904145fcceef79f11b1da9bcf3829edb5055d8a0 of PR #306 the
original version of our t/t-happy-path.sh example test script was added,
with an introductory comment describing the one test that was in the file
at the time as a sample test of the Git LFS client.

We have since copied that comment into several other test scripts,
likely by accident, so we remove those comments now.

We also move the comment in the t/t-happy-path.sh script so it directly
precedes the original "happy path" test, as the script now contains a
number of other tests that are not generic tests intended to be used
as examples.