git/git_test.go,t: add bare and tag worktree tests
The GetAllWorkTrees() function in our "git" package was introduced as
part of the implementation of the "git lfs prune" command in PR #742.
This function examines the contents of a repository's ".git/worktrees"
folder and returns an array of Worktree structures for any linked
working trees it finds, plus one Worktree structure for the main
working tree, if one exists.  The "git lfs prune" command then checks
these working trees' current Git references and indexes for any Git LFS
objects it should retain in the Git LFS object cache.

In subsequent commits in this PR we expect to add an alternative
implementation of the GetAllWorkTrees() function which uses the
"git worktree list" command, as this will ensure we remain compatible
with some forthcoming changes in Git that will alter the contents
of the "gitdir" files in ".git/worktrees" hierarchy.  We will need
to retain the legacy implementation of the GetAllWorkTrees() function
as well, as prior to Git version 2.36.0 the "git worktree list" command
is either not available or does not support the full set of options we
will require.

Before introducing our new version of the GetAllWorkTrees() function,
we first expand the set of tests and checks we perform against it
and the "git lfs prune" command.  This will allow us to verify that
both implementations return identical data under the same conditions.

In particular, in our existing TestWorkTrees() Go test function we add
a linked working tree which references a detached HEAD in the form of
local tag instead of a branch, and confirm that its SHA is also used
as its name, and that its type is returned as RefTypeOther.

We also make use of the NoError() assertion to check the returned
error value from the GetAllWorkTrees() function.

As well, we add a new TestWorktreesBareRepo() Go test function which
checks that when GetAllWorkTrees() is run in a bare repository it
does not return a Worktree structure at all, as there is no main
working tree.  To create the bare repository we add a NewBareRepo()
wrapper function to our "util" test utility package.

In our t/t-prune-worktree.sh shell test script, we slightly expand
the set of checks performed by our "prune worktree" test, just to
clarify which objects are pruned by our "git lfs prune" command when
a linked working tree has been removed but the "git worktree prune"
command has not yet been run.

Finally, we add a "prune worktree (bare main)" test to our test
script which confirms that when the "git lfs prune" command is
run in a working tree linked to a bare repository, no Git LFS objects
are pruned, even if they match no retention configuration settings.
This is because the bare main repository has no Git remotes, so
all Git LFS objects are treated as never having been pushed to a
remote and are therefore all ineligible for pruning.