commands,git: rename GetAllWorktrees() function
The GetAllWorkTrees() function in our "git" package was introduced as
part of the implementation of the "git lfs prune" command in PR #742,
along with the Worktree structure and various related functions, such
as the pruneTaskGetRetainedWorktree() function run by the "git lfs prune"
command.

In all instances, we use the terms "Worktree" and "worktree" rather
than the camel-case "WorkTree" or "workTree", except in the name of
the GetAllWorkTrees() function and its associated TestWorkTrees() test
function.

Before we add another implementation of the GetAllWorkTrees() function
in a subsequent commit in this PR, we first revise the function's name
to match that of the structures it returns, which slightly improves the
consistency of our naming scheme.