git/git.go: read worktrees with list command
The GetAllWorktrees() function in our "git" package was introduced as
part of the implementation of the "git lfs prune" command in PR #742.
This function examines the contents of a repository's ".git/worktrees"
folder and returns an array of Worktree structures for any linked
working trees it finds, plus one Worktree structure for the main
working tree, if one exists.  The "git lfs prune" command then checks
these working trees' current Git references and indexes for any Git LFS
objects it should retain in the Git LFS object cache.

As noted in previous commits in this PR, to check for Git LFS objects
which should be retained, we run the "git diff-index" command in each
extant linked working tree.  Specifically, the commands are started by
the DiffIndex() function in our "git" package, which is run by the
NewDiffIndexScanner() function called by the internal functions invoked
by the ScanIndex() method of the GitScanner structure in our "lfs"
package.  That method is run by the pruneTaskGetRetainedIndex() function
of our "git lfs prune" command.

In order to execute the "git diff-index" command in each linked working
tree, we require the full, absolute paths of the working trees.  At the
moment, our GetAllWorktrees() function expects that it can retrieve these
full paths from the "gitdir" files in each working tree's subdirectory
under the ".git/worktrees" directory.

However, forthcoming changes in Git, likely to be released in Git
version 2.48.0, will alter the contents of these "gitdir" files to
contain relative paths.  As a result, we will no longer be able to
rely on them to determine a full path to each linked working tree.
In particular, commit git/git@717af916cd69d2565aa2a7b7d73d895aa92ff392
causes Git to now write relative paths into the "gitdir" files,
while still supporting any legacy absolute paths found in existing
"gitdir" files.

We can avoid the need to try to parse these relative paths and
determine the correct full path by using the "git worktree list"
command to retrieve the list of linked working trees, plus the
main working tree, if any.  This command returns full paths along
with the current Git references associated with each working tree,
plus several additional attributes.

Prior to Git version 2.36.0, though, the "git worktree list" command
either did not exist, lacked support for the "--porcelain" option,
or lacked support for the "-z" option, so we continue to use our
existing GetAllWorktrees() function's implementation when the
available Git version is lower than 2.36.0.  We therefore rename
the existing version of our GetAllWorktrees() function to
getAllWorktreesFromGitDir(), and invoke it at the top of our new
version of the GetAllWorktrees() function if the Git version is
not 2.36.0 or above.

Our new implementation of the GetAllWorktrees() function simply
calls the "git worktree list --porcelain -z" command and parses
the output, creating the same array of Worktree structures as if
we had parsed the contents of the ".git/worktrees" hierarchy
directly, except that we can depend on the Git command to resolve
any relative paths in "gitdir" files into absolute ones for us.

We do need to call the Clean() function of the "path/filepath" package
to ensure the directory paths are reformatted to match those returned
by the Dir() function of the same package, since that is used by our
legacy GetAllWorktrees() function.  On Windows in particular the
Clean() function will adjust the directory path separators to
align with the default for the OS.

Our parsing logic requires that we find at least an initial "worktree"
line and a subsequent "HEAD" line in the output from the "git worktree
list" command.  Assuming those are found and have non-empty values,
we can return a Worktree structure containing them.  However, if
we also find a "bare" attribute in a subsequent line, we do not
return a Worktree structure for this entry, as we want to ignore
bare main repositories in the same manner as our existing
GetAllWorktrees() function does.

Conveniently, the "git worktree list" command checks if the paths it
returns actually exist, and if they do not, it includes a "prunable"
attribute in its output.  If we detect this attribute we then set the
Prunable flag that we introduced into our Worktree structure in a prior
commit in this PR.