t/cmd: clarify case changes for single bytes
In prior commits in this PR we addressed a number of problems in our
test suite in order to fully enable all the tests in our t/t-clone.sh
test script, which previously did not run to completion, as described
in #5658.

In one commit we updated our git-credential-lfstest test utility program
so it performs a simple Windows-to-Unix path translation on the path
provided by Git when it makes a credential helper request to the program.
In the case of the "clone ClientCert" test in our t/t-clone.sh test
script, this path is the one we set in the "http.<url>.sslCert" Git
configuration value, and in the Git Bash environment we use in our
Windows CI jobs, the MSYS2 environment has converted this path into a
Windows-style one.  However, in order to locate the appropriate
credential record file, our test credential helper requires a path that
matches the Unix-style one we used when setting up those record files.

We therefore adjusted the git-credential-lfstest program to simply look
for a leading uppercase Windows volume identifier like "D:" and convert
it into a lowercase version with a leading slash character like "/d".

In another commit we revised our lfstest-gitserver test utility program
so that after it generates the encrypted TLS certificate key that is
used in our "clone ClientCert" test, it converts the DEK-Info header
of the certificate key to use uppercase hexadecimal digits rather
than lowercase ones.  This resolves a problem in which the test fails
on our Debian build target platforms (Debian 10, 11, and 12) because
the GnuTLS library used by libcurl and therefore also by Git is older
than v3.8.0 and lacks support for lowercase hexadecimal strings in
its parsing of DEK-Info headers.

As suggested by larsxschneider in PR review, we can clarify the intent
of both of these sections of code where we convert bytes from lowercase
to uppercase, or vice versa, by using the standard Go library functions
to perform the conversions, so we do that now.