t/t-batch-transfer-size.sh: remove byte-order mark
The t/t-batch-transfer-size.sh test script added in a prior commit
in this PR contains a leading Unicode byte-order mark which causes
the Perl prove(1) command used by our t/Makefile's recipes to be
unable to determine if it should be run by its
TAP::Parser::SourceHandler::Perl module (which looks for Perl scripts)
or its TAP::Parser::SourceHandler::Executable (which looks for
executable files).

We therefore just remove the BOM, and also add a bit of extra
whitespace to the "batch storage download with small batch size" test
to align it with the "batch storage upload with small batch size" test.