t/t-batch-transfer.sh: check SSH session counts
In subsequent commits in this PR we expect to resolve a pair of issues
which prevent us from testing the SSH object transfer protocol with
more than a single object.  The SSH transfer protocol was introduced in
PR #4446, and in commit 691de511613c1c0dbe6ec92dec649b978ce03cf5 of
that PR, the "batch transfers with ssh endpoint (git-lfs-transfer)" test
in our t/t-batch-transfer.sh test script was added to validate that the
new protocol worked as expected.

This test pushes and then fetches a single object, so the issues which
arise when handling multiple objects in a batch do not cause the test to
fail.  Hence we intend to add two other tests to accompany the existing
one so as to validate the SSH transfer protocol with multiple objects.

All these tests will perform object transfers over SSH, but with varying
numbers and types of SSH sessions.  While our existing test only pushes
and fetches a single object, the tests we will add in a later commit
in this PR will push and fetch multiple objects in each batch.

By default, the Begin() method of the adapterBase structure in our "tq"
package starts eight goroutines to process the objects in a batch, each
running the structure's worker() method.  The first worker routine must
start an SSH session, and depending on how quickly other workers pull
objects from the batch transfer queue, additional SSH sessions may be
created as well, one per active worker.

On platforms other than Windows, we set the default value of the
"lfs.ssh.autoMultiplex" configuration option to "true".  When this
option is "true" and the available SSH client is considered to be
compatible with OpenSSH, we attempt use OpenSSH's ControlMaster option
to create a control socket and multiplex all SSH sessions over a
single common connection.  We expect the first SSH session to be
established with a ControlMaster argument value of "yes", and other
sessions with a value of "no".

At present, our existing "batch transfers with ssh endpoint
(git-lfs-transfer)" test does not check the value of the ControlMaster
argument.  As we expect to add tests which may create more than one SSH
session, however, we would like to validate the number of sessions that
create new control sockets, and the number of SSH sessions overall.

To perform these SSH session count checks we add two dedicated
assertion functions, assert_ssh_transfer_sessions() and
assert_ssh_transfer_session_counts(), with the former calling the
latter several times.  These functions are unlikely to be useful
outside of the context of our tests of the SSH object transfer
protocol, so we define them directly in the t/t-batch-transfer.sh
test script rather than in our generic t/testhelpers.sh library.

The primary function, assert_ssh_transfer_sessions(), checks the
number of times the git-lfs-transfer command is executed over SSH,
and confirms that each session has the ControlMaster option set to
"yes", which is always the case for our existing test, so long as we
update the test to force the use of SSH connection multiplexing on
Windows by explicitly setting the "lfs.ssh.autoMultiplex"
configuration option to "true".

The assert_ssh_transfer_sessions() function then uses the secondary
assert_ssh_transfer_session_counts() function to validate that the
expected number of startup, success, and termination trace log messages
are seen for each SSH session.

One specific issue arises in regard to the number of SSH sessions
and git-lfs-transfer commands we expect to see started during a Git
push operation.  At present, the Git LFS client always starts a unique
SSH session to run the git-lfs-transfer command when checking for
Git LFS locks on the objects being pushed, and this session will
create a control socket if the SSH client in use is considered
compatible with OpenSSH and the "lfs.ssh.autoMultiplex" configuration
option is set to "true".  The control socket opened for this session
is distinct from the one created later by the first SSH session
started by the batch transfer worker routines.

The unique SSH session used for the lock verification request during
push operations is managed with a separate instance of the SSHTransfer
structure from our "ssh" package.  At present, the Git LFS client never
calls the setConnectionCount() method with a zero argument on this
instance of the SSHTransfer structure, so the dedicated SSH session
used for lock verification is never explicitly terminated.

Therefore our new assert_ssh_transfer_sessions() function adjusts
the number of git-lfs-transfer command execution messages and the
number of start and success trace log messages it expects when a push
(i.e., "upload") operation is performed to be one higher than the
number of termination trace log messages it expects.  This allows
the function to account for the additional SSH control socket
session started for the lock verification step, and then never
explicitly terminated.