docs/man: note SSH multiplexing default on Windows
Since commit 448b0c4e8814eca1ea2720c665038568d8d2828d in PR #5537
the GetExeAndArgs() function in our "ssh" package sets the default
value of our "lfs.ssh.autoMultiplex" configuration option to "false"
when running on Windows, and "true" otherwise.

This choice was made because the SSH clients available on Windows
may not support multiplexing SSH sessions over a single connection,
as OpenSSH does with its ControlMaster and ControlPath options.

Since some of these SSH clients may fail if they are passed the
ControlMaster and ControlPath options, we require Windows users who
want to use SSH multiplexing to explicitly enable it by setting the
"lfs.ssh.autoMultiplex" option to "true".  See also the discussion in:

  https://github.com/git-lfs/git-lfs/pull/5537#discussion_r1347862427

However, our git-lfs-config(5) manual page was not updated in PR #5537
to reflect the change in the default value of the "lfs.ssh.autoMultiplex"
option on Windows, so we update it now.

Note that users with the Git for Windows project installed will
typically have a version of OpenSSH available which supports the
ControlMaster option.  However, the OpenSSH for Windows client may
not support multiplexing, as noted in PowerShell/Win32-OpenSSH#1328.